มี เพิ่ม ลบ แก้ไข แล้ว เหลือ ค้นหา

+++ routes/contact.js
@@ -3,11 +3,11 @@ var contacts = require('../database/contacts.js');
 router = express.Router();
 
 router.use('/:id', (req, res, next) => {
-    if(req.method === 'POST' && req.body) {
-        req.method = 'put';
+    if(req.method === 'POST' && Object.keys(req.body).length === 0 && req.body.constructor === Object) {
+        req.method = 'delete';
     }
     else if(req.method === 'POST') {
-        req.method = 'delete';
+        req.method = 'put';
     }
     next();
 });
@@ -30,7 +30,26 @@ router.post('/', (req, res) => {
 });
 
 router.put('/:id', (req, res) => {
-    console.log('edit');
+    if(
+        !req.body.edit_firstn ||
+        !req.body.edit_lastn ||
+        !req.body.edit_phone ||
+        !req.body.edit_email
+    ) {
+        res.status(404).send('Not Found.');
+    }
+    else {
+        for(var i = 0; i < contacts.length; i++) {
+            if(contacts[i].id === parseInt(req.params.id)) {
+                contacts[i].first_name = req.body.edit_firstn;
+                contacts[i].last_name = req.body.edit_lastn;
+                contacts[i].phone = req.body.edit_phone;
+                contacts[i].email = req.body.edit_email;
+                break;
+            }
+        }
+        res.redirect('/content');
+    }
 });
 
 router.delete('/:id', (req, res) => {
